// apps/web/src/shared/api/schemas/error-response.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Платформа подбора роботизированных решений API
 * Контракт API авторизации платформы.
 * OpenAPI spec version: 0.1.0
 */
import type { ErrorResponseDetails } from './error-response-details';

export interface ErrorResponse {
  /** Машинно-читаемый код ошибки. */
  code: string;
  /** Сообщение для отображения пользователю. */
  message: string;
  /** Дополнительные данные об ошибке, если они нужны клиенту. */
  details?: ErrorResponseDetails;
}